Move test utilities to separate file
